Split _local_utils into functional modules
The _local_utils module was misnamed as the utilities it contained
may be used outside of zaza proper. For example in mojo.

This change separates the utilities into functional modules:

 cli_utils: For utilities pertaining to running scripts from command line.
 juju_utils: For utilities pertaining to juju commands.
 generic_utils: For helper utilities that are generic in nature.
