+++ unit_tests/utilitites/test_zaza_utilitites_juju_utils.py
@@ -1,0 +1,157 @@
+import mock
+import unit_tests.utils as ut_utils
+from zaza.utilities import juju_utils
+
+
+class TestJujuUtils(ut_utils.BaseTestCase):
+
+    def setUp(self):
+        super(TestJujuUtils, self).setUp()
+
+        # Juju Status Object and data
+        self.key = "instance-id"
+        self.key_data = "machine-uuid"
+        self.machine = "1"
+        self.machine_data = {self.key: self.key_data}
+        self.unit = "app/1"
+        self.unit_data = {"machine": self.machine}
+        self.application = "app"
+        self.application_data = {"units": {self.unit: self.unit_data}}
+        self.subordinate_application = "subordinate_application"
+        self.subordinate_application_data = {
+            "subordinate-to": [self.application]}
+        self.juju_status = mock.MagicMock()
+        self.juju_status.name = "juju_status_object"
+        self.juju_status.applications.get.return_value = self.application_data
+        self.juju_status.machines.get.return_value = self.machine_data
+
+        # Model
+        self.patch_object(juju_utils, "model")
+        self.patch_object(juju_utils.lifecycle_utils, "get_juju_model")
+        self.model_name = "model-name"
+        self.get_juju_model.return_value = self.model_name
+        self.model.get_status.return_value = self.juju_status
+        self.run_output = {"Code": "0", "Stderr": "", "Stdout": "RESULT"}
+        self.error_run_output = {"Code": "1", "Stderr": "ERROR", "Stdout": ""}
+        self.model.run_on_unit.return_value = self.run_output
+
+        # Clouds
+        self.cloud_name = "FakeCloudName"
+        self.cloud_type = "FakeCloudType"
+        self.clouds = {
+            "clouds":
+                {self.cloud_name:
+                    {"type": self.cloud_type}}}
+
+        # Controller
+        self.patch_object(juju_utils, "controller")
+        self.controller.get_cloud.return_value = self.cloud_name
+
+    def test_get_application_status(self):
+        self.patch_object(juju_utils, "get_full_juju_status")
+        self.get_full_juju_status.return_value = self.juju_status
+
+        # Full status juju object return
+        self.assertEqual(
+            juju_utils.get_application_status(), self.juju_status)
+        self.get_full_juju_status.assert_called_once()
+
+        # Application only dictionary return
+        self.assertEqual(
+            juju_utils.get_application_status(application=self.application),
+            self.application_data)
+
+        # Unit no application dictionary return
+        self.assertEqual(
+            juju_utils.get_application_status(unit=self.unit),
+            self.unit_data)
+
+    def test_get_cloud_configs(self):
+        self.patch_object(juju_utils.Path, "home")
+        self.patch_object(juju_utils.generic_utils, "get_yaml_config")
+        self.get_yaml_config.return_value = self.clouds
+
+        # All the cloud configs
+        self.assertEqual(juju_utils.get_cloud_configs(), self.clouds)
+
+        # With cloud specified
+        self.assertEqual(juju_utils.get_cloud_configs(self.cloud_name),
+                         self.clouds["clouds"][self.cloud_name])
+
+    def test_get_full_juju_status(self):
+        self.assertEqual(juju_utils.get_full_juju_status(), self.juju_status)
+        self.model.get_status.assert_called_once_with(self.model_name)
+
+    def test_get_machines_for_application(self):
+        self.patch_object(juju_utils, "get_application_status")
+        self.get_application_status.return_value = self.application_data
+
+        # Machine data
+        self.assertEqual(
+            juju_utils.get_machines_for_application(self.application),
+            [self.machine])
+        self.get_application_status.assert_called_once()
+
+        # Subordinate application has no units
+        def _get_application_status(application):
+            _apps = {
+                self.application: self.application_data,
+                self.subordinate_application:
+                    self.subordinate_application_data}
+            return _apps[application]
+        self.get_application_status.side_effect = _get_application_status
+
+        self.assertEqual(
+            juju_utils.get_machines_for_application(
+                self.subordinate_application),
+            [self.machine])
+
+    def test_get_machine_status(self):
+        self.patch_object(juju_utils, "get_full_juju_status")
+        self.get_full_juju_status.return_value = self.juju_status
+
+        # All machine data
+        self.assertEqual(
+            juju_utils.get_machine_status(self.machine),
+            self.machine_data)
+        self.get_full_juju_status.assert_called_once()
+
+        # Request a specific key
+        self.assertEqual(
+            juju_utils.get_machine_status(self.machine, self.key),
+            self.key_data)
+
+    def test_get_machine_uuids_for_application(self):
+        self.patch_object(juju_utils, "get_machines_for_application")
+        self.get_machines_for_application.return_value = [self.machine]
+
+        self.assertEqual(
+            juju_utils.get_machine_uuids_for_application(self.application),
+            [self.machine_data.get("instance-id")])
+        self.get_machines_for_application.assert_called_once_with(
+            self.application)
+
+    def test_get_provider_type(self):
+        self.patch_object(juju_utils, "get_cloud_configs")
+        self.get_cloud_configs.return_value = {"type": self.cloud_type}
+        self.assertEqual(juju_utils.get_provider_type(),
+                         self.cloud_type)
+        self.get_cloud_configs.assert_called_once_with(self.cloud_name)
+
+    def test_remote_run(self):
+        _cmd = "do the thing"
+
+        # Success
+        self.assertEqual(juju_utils.remote_run(self.unit, _cmd),
+                         self.run_output["Stdout"])
+        self.model.run_on_unit.assert_called_once_with(
+            self.model_name, self.unit, _cmd, timeout=None)
+
+        # Non-fatal failure
+        self.model.run_on_unit.return_value = self.error_run_output
+        self.assertEqual(juju_utils.remote_run(self.unit, _cmd, fatal=False),
+                         self.error_run_output["Stderr"])
+
+        # Fatal failure
+        with self.assertRaises(Exception):
+            juju_utils.remote_run(self.unit, _cmd, fatal=True)
