+++ zaza/charm_tests/dragent/tests.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from zaza.utilities import _local_utils
+from zaza.utilities import generic_utils
 from zaza.charm_tests.dragent import test
 
 
@@ -12,7 +12,7 @@ class DRAgentTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        _local_utils.setup_logging()
+        generic_utils.setup_logging()
 
     def test_bgp_routes(self):
         test.test_bgp_routes(peer_application_name=self.BGP_PEER_APPLICATION)
